src/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
astroid==3.0.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.1.0
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
databases==0.8.0
dill==0.3.7
fastapi==0.104.1
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.1
idna==3.4
iniconfig==2.0.0
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
more-itertools==10.1.0
packaging==23.2
platformdirs==4.0.0
pluggy==1.3.0
psycopg2-binary==2.9.9
py==1.11.0
pydantic==2.5.1
pydantic_core==2.14.3
pylint==3.0.2
pyparsing==3.1.1
pytest==7.4.3
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.50
starlette==0.27.0
tomlkit==0.12.3
typing_extensions==4.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
uvloop==0.19.0
wcwidth==0.2.12
websockets==12.0
wrapt==1.16.0
prometheus-fastapi-instrumentator
prometheus-client
debugpy